Add Hammerspoon
Quoting from https://www.hammerspoon.org/

> What is Hammerspoon?
> This is a tool for powerful automation of OS X. At its core, Hammerspoon
> is just a bridge between the operating system and a Lua scripting
> engine. What gives Hammerspoon its power is a set of extensions that
> expose specific pieces of system functionality, to the user.

Since it is a configuration based tool for automation, it's preferred
over any other tool with similar functionality.

At the moment I am using this for adding a global hotkey to open
/ switch to Alacritty to simulate similar behavior to iTerm2's hotkey
that I became habbitual to. I opted for <C-`> as the hotkey to keep it
a bit more sane.